created useModal hook to isolate modal open/close functionality. Used the hook in Drawer and Modal components
