+++ src/shared/lib/hooks/useModal/useModal.tsx
@@ -1,0 +1,69 @@
+import { useTheme } from 'app/providers/ThemeProvider';
+import {
+    useState,
+    useRef,
+    MutableRefObject,
+    useCallback,
+    useEffect,
+} from 'react';
+
+interface UseModalProps {
+    onClose?: () => void;
+    isOpen?: boolean;
+    animationDelay: number;
+}
+export function useModal(props: UseModalProps) {
+    const { animationDelay, isOpen, onClose } = props;
+
+    const [isClosing, setIsClosing] = useState(false);
+    const [isMounted, setIsMounted] = useState(false);
+
+    const timerRef = useRef() as MutableRefObject<
+        ReturnType<typeof setTimeout>
+    >;
+
+    const close = useCallback(() => {
+        if (onClose) {
+            setIsClosing(true);
+            timerRef.current = setTimeout(() => {
+                onClose();
+                setIsClosing(false);
+            }, animationDelay);
+        }
+    }, [animationDelay, onClose]);
+
+    //on every rerender of the component, the the reference to the onKeyDown will be new if not wrapped in useCallback
+    const onKeyDown = useCallback(
+        (e: KeyboardEvent) => {
+            if (e.key === 'Escape') {
+                close();
+            }
+        },
+        [close]
+    );
+    // const onContentClick = (e: React.MouseEvent) => {
+    //     e.stopPropagation();
+    // };
+
+    useEffect(() => {
+        if (isOpen) {
+            setIsMounted(true);
+        }
+    }, [isOpen]);
+
+    useEffect(() => {
+        if (isOpen) {
+            window.addEventListener('keydown', onKeyDown);
+        }
+        return () => {
+            clearTimeout(timerRef.current);
+            window.removeEventListener('keydown', onKeyDown);
+        };
+    }, [isOpen, onKeyDown]);
+
+    return {
+        isClosing,
+        isMounted,
+        close,
+    };
+}
