+++ src/shared/ui/Modal/Modal.tsx
@@ -1,18 +1,10 @@
 import { useTheme } from 'app/providers/ThemeProvider';
-import {
-    FC,
-    MutableRefObject,
-    ReactNode,
-    useCallback,
-    useEffect,
-    useRef,
-    useState,
-} from 'react';
+import { FC, ReactNode } from 'react';
 import { classNames, Modes } from 'shared/lib/classnames/classNames';
 import { Portal } from '../Portal/Portal';
 import cls from './Modal.module.scss';
 import { Overlay } from '../Overlay/Overlay';
-
+import { useModal } from 'shared/lib/hooks/useModal/useModal';
 interface ModalProps {
     className?: string;
     children?: ReactNode;
@@ -20,57 +12,18 @@ interface ModalProps {
     onClose?: () => void;
     lazy?: boolean;
 }
-const ANIMATION_DELAY = 300;
+
 
 export const Modal: FC<ModalProps> = (props: ModalProps) => {
     const { className, children, isOpen, onClose, lazy } = props;
 
-    const [isClosing, setIsClosing] = useState(false);
-    const [isMounted, setIsMounted] = useState(false);
-
-    const timerRef = useRef() as MutableRefObject<
-        ReturnType<typeof setTimeout>
-    >;
-
     const { theme } = useTheme();
-    const closeHandler = useCallback(() => {
-        if (onClose) {
-            setIsClosing(true);
-            timerRef.current = setTimeout(() => {
-                onClose();
-                setIsClosing(false);
-            }, ANIMATION_DELAY);
-        }
-    }, [onClose]);
 
-    //on every rerender of the component, the the reference to the onKeyDown will be new if not wrapped in useCallback
-    const onKeyDown = useCallback(
-        (e: KeyboardEvent) => {
-            if (e.key === 'Escape') {
-                closeHandler();
-            }
-        },
-        [closeHandler]
-    );
-    const onContentClick = (e: React.MouseEvent) => {
-        e.stopPropagation();
-    };
-
-    useEffect(() => {
-        if (isOpen) {
-            setIsMounted(true);
-        }
-    }, [isOpen]);
-
-    useEffect(() => {
-        if (isOpen) {
-            window.addEventListener('keydown', onKeyDown);
-        }
-        return () => {
-            clearTimeout(timerRef.current);
-            window.removeEventListener('keydown', onKeyDown);
-        };
-    }, [isOpen, onKeyDown]);
+    const { close, isClosing, isMounted } = useModal({
+        animationDelay: 300,
+        onClose,
+        isOpen,
+    });
 
     const modes: Modes = {
         [cls.opened]: isOpen,
@@ -83,11 +36,15 @@ export const Modal: FC<ModalProps> = (props: ModalProps) => {
 
     return (
         <Portal>
-            <div className={classNames(cls.Modal, modes, [className, theme, 'app_modal'])}>
-                <Overlay onClick={closeHandler}>
-                    <div className={cls.content} >
-                        {children}
-                    </div>
+            <div
+                className={classNames(cls.Modal, modes, [
+                    className,
+                    theme,
+                    'app_modal',
+                ])}
+            >
+                <Overlay onClick={close}>
+                    <div className={cls.content}>{children}</div>
                 </Overlay>
             </div>
         </Portal>
